+++ src/ts/localization/i18n.ts
@@ -129,6 +129,8 @@ export interface Vocabulary {
   'contextMenu.subtitle': string;
   ecoMode: string;
   'ecoMode.title': string;
+  persistentPreferences: string;
+  'persistentPreferences.title': string;
 }
 
 /**